image_process/azure-cowrec-func/app_functions/requirements.txt
azure-functions
tensorflow==1.14.0
pillow==6.1.0
numpy==1.16.4
requests